pyproject.toml
[tool.poetry]
name = "nDIDI"
version = "1.0.0b"
description = ""
authors = ["LEv145 <levxxx03@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.11"
hikari-tanjun = "^2.4.0-alpha.1"
hikari = "^2.0.0-alpha.106"

[tool.poetry.dev-dependencies]
mkinit = "^0.3.4"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
